bootstraped a default admin user

+++ src/main/java/com/example/knot/service/UserService.java
@@ -4,10 +4,11 @@ import com.example.knot.dto.UpdateUserRequest;
 import com.example.knot.dto.UserResponse;
 import com.example.knot.entity.User;
 import com.example.knot.exception.AlreadyFollowingException;
-import com.example.knot.exception.EmailAlreadyExistsException;
 import com.example.knot.exception.NotFollowingException;
+import com.example.knot.exception.UnauthorizedActionException;
 import com.example.knot.exception.UserNotFoundException;
 import com.example.knot.repository.UserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,6 +23,22 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
+
+    public UserResponse getCurrentUser() {
+        UUID userId = (UUID) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found"));
+        return UserResponse.builder()
+                .id(user.getId())
+                .name(user.getName())
+                .email(user.getEmail())
+                .bio(user.getBio())
+                .createdAt(user.getCreatedAt())
+                .build();
+    }
+
+    public UUID getCurrentUserId() {
+        return (UUID) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    }
 
     public UserResponse getUserById(UUID id) {
         User user = userRepository.findById(id).orElseThrow(()->new UserNotFoundException("User Not Found"));
@@ -50,6 +67,12 @@ public class UserService {
 
     public UserResponse updateUser(UUID id, UpdateUserRequest request) {
 
+        UUID userId = getCurrentUserId();
+
+        if(!userId.equals(id)) {
+            throw new UnauthorizedActionException("Cant update others profiles");
+        }
+
         User user = userRepository.findById(id).orElseThrow(()->new UserNotFoundException("User Not Found"));
         user.setBio(request.getBio());
         user.setName(request.getName());
@@ -65,6 +88,12 @@ public class UserService {
     }
 
     public void deleteUser(UUID id) {
+
+        UUID userId = getCurrentUserId();
+        if(!userId.equals(id)) {
+            throw new UnauthorizedActionException("Cant delete others profiles");
+        }
+
         User user = userRepository.findById(id)
                 .orElseThrow(()->new UserNotFoundException("User Not Found"));
         userRepository.delete(user);
@@ -72,7 +101,7 @@ public class UserService {
 
     public void followUser(UUID userId, UUID targetId) {
         if(userId.equals(targetId)) {
-            throw new RuntimeException("You can't follow yourself");
+            throw new AlreadyFollowingException("You can't follow yourself");
         }
         User user = userRepository.findById(userId)
                 .orElseThrow(()->new UserNotFoundException("User Not Found"));
